requirements.txt
anyio==4.13.0
APScheduler==3.10.4
asgiref==3.11.1
blinker==1.9.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
reportlab==4.0.9
cryptography==46.0.6
Django==5.2.6
django-apscheduler==0.7.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
firebase_admin==7.3.0
Flask==3.1.3
google-api-core==2.30.1
google-auth==2.49.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
pillow==12.2.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pytz==2026.1.post1
PyYAML==6.0.1
razorpay==2.0.1
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.8
whitenoise==6.12.0
